OSIS-4817 : Ajout de l'action "Liste des formations" dans la page de recherche des of + Suppression dans l'ancienne view
