+++ program_management/views/search.py
@@ -6,7 +6,7 @@
 #    The core business involves the administration of students, teachers,
 #    courses, programs and so on.
 #
-#    Copyright (C) 2015-2019 Université catholique de Louvain (http://www.uclouvain.be)
+#    Copyright (C) 2015-2020 Université catholique de Louvain (http://www.uclouvain.be)
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -34,13 +34,14 @@ from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 from django_filters.views import FilterView
 
+from base.business.education_group import create_xls, ORDER_COL, ORDER_DIRECTION
 from base.forms.search.search_form import get_research_criteria
 from base.models.academic_year import starting_academic_year
 from base.models.education_group_type import EducationGroupType
 from base.models.enums import education_group_categories
 from base.models.person import Person
 from base.utils.cache import CacheFilterMixin
-from base.utils.search import SearchMixin
+from base.utils.search import SearchMixin, RenderToExcel
 from education_group.models.group_year import GroupYear
 from program_management.api.serializers.education_group import EducationGroupSerializer
 from program_management.forms.education_groups import GroupFilter
@@ -50,6 +51,17 @@ def _get_filter(form):
     return OrderedDict(itertools.chain(get_research_criteria(form)))
 
 
+def _create_xls(view_obj, context, **response_kwargs):
+    user = view_obj.request.user
+    egys = context["filter"].qs
+    filters = _get_filter(context["form"])
+    # FIXME: use ordering args in filter_form! Remove xls_order_col/xls_order property
+    order = {ORDER_COL: view_obj.request.GET.get('xls_order_col'),
+             ORDER_DIRECTION: view_obj.request.GET.get('xls_order')}
+    return create_xls(user, egys, filters, order)
+
+
+@RenderToExcel("xls", _create_xls)
 class EducationGroupSearch(LoginRequiredMixin, PermissionRequiredMixin, CacheFilterMixin, SearchMixin, FilterView):
     model = GroupYear
     template_name = "search.html"
